카테고리 정보 추가

+++ src/main/java/com/repill/was/member/entity/member/MemberRepository.java
@@ -20,5 +20,5 @@ public interface MemberRepository {
 
     Optional<Member> findByMemberNickName(String nickName);
 
-    Optional<Member> getFavoriteItems(MemberId memberId, int size, int cursorId);
+    Optional<Member> getFavoriteItems(MemberId memberId, int size, Long cursorId);
 }
